Refactored manager pages into new filesystem
